gateway/requirements.txt
flask==2.3.3
flask-cors==4.0.2
grpcio==1.60.0
protobuf==4.25.1
werkzeug==3.0.6
gunicorn==22.0.0